pyproject.toml
[project]
name = "internvl-finetune"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain[openai]>=0.3.25",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
